requirements.txt
numpy==1.24.4
pandas==2.0.3
scikit-learn==1.3.2
xgboost==2.1.4
pickle5==0.0.11

azureml-mlflow==1.56.0
azureml-core==1.55.0

mlflow==2.12.1
joblib==1.3.2